# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Societal Innovation Portal API
  version: 1.0.0
  description: >
    Canonical machine-readable contract for the SIH 2026 Societal Innovation Portal.
    Human-facing notes, examples and socket events live in docs/API_CONTRACT.md.
    Request/response shapes are generated from packages/shared-types (Zod) --
    if the two disagree, packages/shared-types wins and this file is the bug.
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://report-maro-c8m1.onrender.com
    description: Deployed demo
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/problems/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Problems]
      summary: Problem detail including AI analysis
      security: []
      responses:
        '200':
          description: Problem
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Problem' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/assignments:
    get:
      tags: [Assignments]
      summary: List assignments (ADMIN sees all, UNIVERSITY sees its own)
      parameters:
        - { in: query, name: status, schema: { $ref: '#/components/schemas/AssignmentStatus' } }
        - { in: query, name: universityId, schema: { type: string } }
        - { in: query, name: problemId, schema: { type: string } }
        - { in: query, name: mine, schema: { type: boolean } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: Paginated assignments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Assignment' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
    post:
      tags: [Assignments]
      summary: Assign a problem to a university (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [problemId, universityId]
              properties:
                problemId: { type: string }
                universityId: { type: string }
                note: { type: string, maxLength: 1000 }
                dueDate: { type: string, format: date-time }
      responses:
        '201':
          description: Assignment created; problem moves to ASSIGNED
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Assignment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/assignments/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Assignments]
      summary: Assignment detail
      responses:
        '200':
          description: Assignment
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Assignment' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Assignments]
      summary: Accept, decline or complete an assignment (UNIVERSITY)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [ACCEPTED, DECLINED, COMPLETED] }
                reason: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Assignment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: The caller's notification feed
      parameters:
        - { in: query, name: unreadOnly, schema: { type: boolean } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: Paginated notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Notification' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
...
  /api/v1/notifications/{id}/read:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      tags: [Notifications]
      summary: Mark one notification read
      responses:
        '200':
          description: Updated notification
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Notification' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/internal/problems/search-similar:
    post:
      tags: [Internal]
      summary: pgvector nearest-neighbour lookup for dedupe
      security:
        - internalKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [embedding]
              properties:
                embedding:
                  type: array
                  items: { type: number }
                  minItems: 384
                  maxItems: 384
                excludeProblemId: { type: string }
                district: { type: string }
                limit: { type: integer, maximum: 20, default: 5 }
                minScore: { type: number, default: 0.8 }
      responses:
        '200':
          description: Candidate duplicates ordered by descending cosine similarity
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: object
                    properties:
                      candidates:
                        type: array
                        items: { $ref: '#/components/schemas/SimilarityCandidate' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/internal/problems/{id}/analysis:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      tags: [Internal]
      summary: Write AI results back (idempotent)
      security:
        - internalKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnalysisCallbackRequest' }
      responses:
        '200':
          description: Applied (or replayed, with applied=false)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: object
                    properties:
                      problemId: { type: string }
                      status: { $ref: '#/components/schemas/ProblemStatus' }
                      applied: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/internal/problems/{id}/failure:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      tags: [Internal]
      summary: Report a permanently failed analysis
      description: Called after BullMQ exhausts retries so the problem does not sit in PROCESSING forever.
      security:
        - internalKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, attempts]
              properties:
                reason: { type: string }
                attempts: { type: integer }
                stack: { type: string }
      responses:
        '202': { description: Recorded; problem returned to SUBMITTED for manual triage }
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalKey:
      type: apiKey
      in: header
      name: x-internal-key
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Authenticated but the role is not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Illegal state transition or duplicate
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiError:
      type: object
      required: [success, error, requestId]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, PAYLOAD_TOO_LARGE, INTERNAL_ERROR, SERVICE_UNAVAILABLE]
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  path: { type: string }
                  message: { type: string }
        requestId: { type: string }
    ListMeta:
      type: object
      properties:
        pagination:
          type: object
          required: [page, pageSize, total, totalPages]
          properties:
            page: { type: integer }
            pageSize: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    ProblemCategory:
      type: string
      enum: [WATER_SANITATION, ROADS_TRANSPORT, EDUCATION, HEALTHCARE, ELECTRICITY, WASTE_MANAGEMENT, AGRICULTURE, PUBLIC_SAFETY, OTHER]
    Priority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    ProblemStatus:
      type: string
      enum: [SUBMITTED, PROCESSING, TRIAGED, DUPLICATE, ASSIGNED, IN_PROGRESS, RESOLVED, REJECTED]
    AssignmentStatus: { type: string, enum: [PENDING, ACCEPTED, DECLINED, COMPLETED] }
...
    OrganizationSummary:
      type: object
      required: [id, name, type, district]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, enum: [UNIVERSITY, INDUSTRY, GOVERNMENT] }
        district: { type: [string, 'null'] }
...
    Location:
      type: object
      required: [latitude, longitude, district]
      properties:
        latitude: { type: number, minimum: 21.5, maximum: 25.6 }
        longitude: { type: number, minimum: 83.2, maximum: 88.0 }
        district: { type: string }
        address: { type: string }
    Photo:
      type: object
      required: [key, url]
      properties:
        key: { type: string }
        url: { type: string, format: uri }
    ProblemAnalysis:
      type: object
      required: [category, categoryConfidence, priority, priorityScore, keywords, duplicateOfId, similarityScore, model, analyzedAt]
      properties:
        category: { $ref: '#/components/schemas/ProblemCategory' }
        categoryConfidence: { type: number, minimum: 0, maximum: 1 }
        priority: { $ref: '#/components/schemas/Priority' }
        priorityScore: { type: number, minimum: 0, maximum: 100 }
        keywords: { type: array, items: { type: string } }
        duplicateOfId: { type: [string, 'null'] }
        similarityScore: { type: [number, 'null'] }
        model: { type: string }
        analyzedAt: { type: string, format: date-time }
    ProblemSummary:
      type: object
      required: [id, title, status, location, reporter, category, priority, duplicateCount, assignmentId, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        status: { $ref: '#/components/schemas/ProblemStatus' }
        location: { $ref: '#/components/schemas/Location' }
        reporter:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        category:
          oneOf:
            - $ref: '#/components/schemas/ProblemCategory'
            - type: 'null'
        priority:
          oneOf:
            - $ref: '#/components/schemas/Priority'
            - type: 'null'
        duplicateCount: { type: integer }
        assignmentId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Problem:
      allOf:
        - $ref: '#/components/schemas/ProblemSummary'
        - type: object
          required: [description, photos, analysis]
          properties:
            description: { type: string }
            photos: { type: array, items: { $ref: '#/components/schemas/Photo' } }
            analysis:
              oneOf:
                - $ref: '#/components/schemas/ProblemAnalysis'
                - type: 'null'
...
    Assignment:
      type: object
      required: [id, problemId, problem, university, status, note, dueDate, assignedBy, respondedAt, proposalId, createdAt, updatedAt]
      properties:
        id: { type: string }
        problemId: { type: string }
        problem: { $ref: '#/components/schemas/ProblemSummary' }
        university: { $ref: '#/components/schemas/OrganizationSummary' }
        status: { $ref: '#/components/schemas/AssignmentStatus' }
        note: { type: [string, 'null'] }
        dueDate: { type: [string, 'null'], format: date-time }
        assignedBy:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        respondedAt: { type: [string, 'null'], format: date-time }
        proposalId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Notification:
      type: object
      required: [id, type, title, body, link, readAt, createdAt]
      properties:
        id: { type: string }
        type:
          type: string
          enum: [PROBLEM_ANALYZED, PROBLEM_ASSIGNED, PROBLEM_STATUS_CHANGED, PROPOSAL_SUBMITTED, PROPOSAL_REVIEWED, PROJECT_FUNDED]
        title: { type: string }
        body: { type: string }
        link: { type: [string, 'null'] }
        readAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
...
    SimilarityCandidate:
      type: object
      properties:
        problemId: { type: string }
        title: { type: string }
        score: { type: number }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    AnalysisCallbackRequest:
      type: object
      required: [category, categoryConfidence, priority, priorityScore, embedding, model, processingMs]
      properties:
        category: { $ref: '#/components/schemas/ProblemCategory' }
        categoryConfidence: { type: number, minimum: 0, maximum: 1 }
        priority: { $ref: '#/components/schemas/Priority' }
        priorityScore: { type: number, minimum: 0, maximum: 100 }
        keywords: { type: array, items: { type: string } }
        embedding:
          type: array
          items: { type: number }
          minItems: 384
          maxItems: 384
        duplicateOfId: { type: [string, 'null'] }
        similarityScore: { type: [number, 'null'] }
        model: { type: string }
        processingMs: { type: integer }